+++ packages/gasket-analyze-plugin/lib/index.js
@@ -1,6 +1,6 @@
-const merge = require('deepmerge');
-const { BundleAnalyzerPlugin } = require('webpack-bundle-analyzer');
-
+const create = require('./create');
+const webpack = require('./webpack');
+const getCommands = require('./get-commands');
 /**
  * Gasket Analyzer Plugin
  *
@@ -9,59 +9,8 @@ const { BundleAnalyzerPlugin } = require('webpack-bundle-analyzer');
 module.exports = {
   name: 'analyze',
   hooks: {
-    /**
-     * Add the analyzer webpack plugin if analyze flag has been set
-     *
-     * @param {Object} gasket - Gasket API
-     * @param {Object} webpackConfig - Webpack config
-     * @param {Object} data - Next.js data
-     * @returns {Object} webpackConfig
-     */
-    webpack(gasket, webpackConfig, data) {
-      const {
-        config: {
-          analyze = false,
-          bundleAnalyzerConfig: userConfig = {}
-        }
-      } = gasket;
-
-      //
-      // analyze will be set by the cli
-      //
-      if (analyze) {
-        const { isServer } = data;
-        const bundleAnalyzerConfig = merge(require('./default-config'),
-          userConfig);
-        const { browser = {}, server = {} } = bundleAnalyzerConfig;
-
-        //
-        // return webpack config partial
-        //
-        return {
-          plugins: [
-            new BundleAnalyzerPlugin({
-              ...(isServer ? server : browser)
-            })]
-        };
-      }
-
-      return null;
-    },
-
-    /**
-     * Add files & extend package.json for new apps.
-     *
-     * @param {Gasket} gasket - The gasket API.
-     * @param {CreateContext} context - Create context
-     * @returns {Promise} promise
-     * @public
-     */
-    async create(gasket, context) {
-      const { pkg } = context;
-
-      pkg.add('scripts', {
-        analyze: 'gasket analyze'
-      });
-    }
+    webpack,
+    getCommands,
+    create
   }
 };