+++ packages/gasket-metrics-plugin/lib/metrics.js
@@ -1,44 +1,23 @@
-const Git = require('git-shizzle');
-const path = require('path');
 const os = require('os');
-const fs = require('fs');
+const Git = require('git-shizzle');
 const gitUrlParse = require('git-url-parse');
 const debug = require('diagnostics')('gasket:cli:metrics');
 
 class Metrics {
   /**
    * Instantiate metrics object tracker
-   * @param  {Object} config gasket config
-   * @param  {Boolean} record whether or not to record the event
-   * @param  {String} cmd what gasket command is being run
+   * @param  {Gasket} gasket Gasket API
+   * @param  {Object} gasket.config gasket config
+   * @param  {Boolean} gasket.record whether or not to record the event
+   * @param  {String} gasket.cmd what gasket command is being run
    */
-  constructor(config = {}, record, cmd) {
+  constructor(gasket) {
+    const { command, config, metadata } = gasket;
     this.gitshizzle = new Git(config.root);
     this.config = config;
-    this.record = record;
-    this.cmd = cmd;
-  }
-
-  /**
-   * Attempt to read the `package.json` that is in the root of the project.
-   *
-   * @returns {Promise} The parsed `package.json`
-   * @public
-   */
-  json() {
-    const location = path.join(this.config.root, 'package.json');
-
-    return new Promise(function promised(resolve) {
-      fs.readFile(location, 'utf-8', function read(err, data) {
-        if (err) return resolve({});
-
-        let parsed;
-
-        try { parsed = JSON.parse(data); } catch (e) { return resolve({}); }
-
-        resolve(parsed);
-      });
-    });
+    this.cmd = command;
+    this.record = command.flags.record;
+    this.metadata = metadata;
   }
 
   /**
@@ -68,11 +47,12 @@ class Metrics {
    * @public
    */
   async collect() {
-    const [remote, branch, packagejson] = await Promise.all([
+    const [remote, branch] = await Promise.all([
       this.git('config', '--get remote.origin.url'),
-      this.git('revParse', '--abbrev-ref HEAD'),
-      this.json()
+      this.git('revParse', '--abbrev-ref HEAD')
     ]);
+
+    const packagejson = this.metadata.app.package;
 
     const config = this.config;
     const dependencies = packagejson.dependencies || {};
@@ -105,7 +85,7 @@ class Metrics {
       argv: process.argv.slice(2),
 
       time: Date.now(),
-      cmd: this.cmd
+      cmd: this.cmd.id
     };
   }