+++ packages/gasket-metrics-plugin/test/metrics.test.js
@@ -1,15 +1,43 @@
 const assume = require('assume');
+const sinon = require('sinon');
 const path = require('path');
 const os = require('os');
-const Metrics = require('../../src/metrics');
+const Metrics = require('../lib/metrics');
+
+const mockPackage = {
+  name: 'mock-app',
+  version: '7.8.9',
+  dependencies: {
+    '@gasket/plugin-engine': '1.2.3',
+    '@gasket/example-plugin': '2.3.4',
+    '@gasket/example-preset': '3.4.5'
+  }
+};
 
 describe('Metrics', function () {
-  const gasket = {
-    root: path.join(__dirname, '..', '..'),
-    env: 'custom-env-set-here'
-  };
+  let mockGasket, metrics;
 
-  const metrics = new Metrics(gasket, true, 'start');
+  beforeEach(() => {
+    mockGasket = {
+      command: {
+        id: 'test-command',
+        flags: {
+          record: true
+        }
+      },
+      config: {
+        root: path.join(__dirname, '..'),
+        env: 'custom-env-set-here'
+      },
+      metadata: {
+        app: {
+          package: mockPackage
+        }
+      }
+    };
+
+    metrics = new Metrics(mockGasket);
+  });
 
   describe('collect', function () {
     it('returns an object with', async function () {
@@ -18,23 +46,15 @@ describe('Metrics', function () {
       assume(data).is.a('object');
     });
 
-    it('works in a directory without git or package.json', async function () {
-      const config = Object.assign({}, gasket);
+    it('works in a directory without git', async function () {
+      mockGasket.config.root = os.tmpdir();
 
-      //
-      // Major assumption ahead. We assume that the parent directory that
-      // holds this repo has no package.json or git repo.
-      //
-      config.root = path.join(config.root, '..', '..', '..');
-
-      const parent = new Metrics(config);
+      const parent = new Metrics(mockGasket);
       const data = await parent.collect();
 
       assume(data).is.a('object');
       assume(data.branch).is.a('undefined');
       assume(data.repository).is.a('undefined');
-      assume(data.name).is.a('undefined');
-      assume(data.version).is.a('undefined');
     });
 
     it('includes git data', async function () {
@@ -49,12 +69,11 @@ describe('Metrics', function () {
       assume(data.repository).includes('gasket.git');
     });
 
-    it('includes package.json data', async function () {
+    it('includes package from app metadata', async function () {
       const data = await metrics.collect();
-      const pkg = require('../../package.json');
 
-      assume(data.name).equals(pkg.name);
-      assume(data.version).equals(pkg.version);
+      assume(data.name).equals(mockPackage.name);
+      assume(data.version).equals(mockPackage.version);
       assume(data.gasket).owns('@gasket/plugin-engine');
     });
 
@@ -69,7 +88,7 @@ describe('Metrics', function () {
     it('includes gasket data', async function () {
       const data = await metrics.collect();
 
-      assume(data.env).equals(gasket.env);
+      assume(data.env).equals(mockGasket.config.env);
 
       //
       // To confirm that we are not accidently including user paths to a
@@ -85,7 +104,7 @@ describe('Metrics', function () {
 
     it('includes the gasket command', async function () {
       const { cmd } = await metrics.collect();
-      assume(cmd).equals('start');
+      assume(cmd).equals('test-command');
     });
 
     it('include the gasket dependencies in a list', async function () {
@@ -96,15 +115,31 @@ describe('Metrics', function () {
   });
 
   describe('report', function () {
+    let collectStub;
+
+    beforeEach(() => {
+      collectStub = sinon.stub(Metrics.prototype, 'collect');
+    });
+
+    afterEach(() => {
+      collectStub.restore();
+    });
 
     it('Returns exactly what is collected', async function () {
       const data = { command: 'let\'s make this matzo' };
+      collectStub.resolves(data);
+      const metricData = await metrics.report();
 
-      metrics.record = true;
-      metrics.collect = async () => data;
-      const metricData = await metrics.report(true);
-
+      assume(collectStub).called();
       assume(metricData).deep.equals(data);
+    });
+
+    it('does not collect if record disabled', async function () {
+      mockGasket.command.flags.record = false;
+      metrics = new Metrics(mockGasket);
+      await metrics.report();
+
+      assume(collectStub).not.called();
     });
   });
 });